requirements.txt
requests
python-dotenv
pytest>=6.0
pytest-mock>=3.0
matplotlib>=3.0